requirements.txt
torch
transformers
Pillow
numpy
matplotlib
tqdm
annoy
